update tests for player info
